Improve Django support
Implement various improvements to the Django middleware, making it
easier to customize or extend to suit more complex use-cases.

Default behavior of the middleware has also been changed to not profile
every request, but instead only if a special header or query parameter
is set. The request filter can also be easily customized.

Trace files created by the middleware are no longer written directly to
the destination. Instead, they get buffered to a temporary file first,
which is then saved to a configurable Django storage backend.